MAS-111 Moved RenderviewAsync frome extension to service
